Moved bulk insert operations into models Base
- As discussed in the PR model layer seems like the most appropriate place for these operations. Follow up will include moving bulk delete operation
